requirements.txt
altair==5.0.1
anyio==3.7.0
attrs==23.1.0
azure-core==1.27.0
blinker==1.6.2
bz2file==0.98
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
decorator==5.1.1
exceptiongroup==1.1.1
fastapi==0.96.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
pyarrow==12.0.0
pybase64==1.2.3
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rich==13.4.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
streamlit==1.23.1
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
